S.P.E.D
Added the S.P.E.D object in the room where upon touching the green godot block, the player wins and resets the numKeys the player has. In the server side, added the remoteGameOver function. Put Player and Creature in the playerDamageable group. Currently door disappears after >3 keys collected for testing purposes since if you run 2 clients on the same godot, it doubles the amount of collisions the player encounters